ATO-177: Move logout functionality to shared
