+++ orchestration-shared/src/main/java/uk/gov/di/orchestration/shared/domain/LogoutAuditableEvent.java
@@ -1,0 +1,10 @@
+package uk.gov.di.orchestration.shared.domain;
+
+@SuppressWarnings("java:S6548")
+public enum LogoutAuditableEvent implements AuditableEvent {
+    LOG_OUT_SUCCESS;
+
+    public AuditableEvent parseFromName(String name) {
+        return valueOf(name);
+    }
+}
